+++ packages/server/test/integration/health.test.ts
@@ -2,7 +2,7 @@ import { HealthEventType, HealthReport } from '@botpress/messaging-base'
 import { Client } from '../../src/clients/types'
 import { Conduit } from '../../src/conduits/types'
 import { HealthService } from '../../src/health/service'
-import { app, randStr, setupApp } from './utils'
+import { app, randStr, setupApp } from '../utils'
 
 const TEST_CHANNEL = 'messenger'
