+++ packages/server/test/integration/providers.test.ts
@@ -1,7 +1,7 @@
 import { validate as validateUuid } from 'uuid'
 import { ProviderService } from '../../src/providers/service'
 import { Provider } from '../../src/providers/types'
-import { app, randStr, setupApp } from './utils'
+import { app, randStr, setupApp } from '../utils'
 
 describe('Providers', () => {
   let providers: ProviderService
