+++ packages/server/test/integration/status.test.ts
@@ -5,7 +5,7 @@ import { Conduit } from '../../src/conduits/types'
 import { Provider } from '../../src/providers/types'
 import { StatusService } from '../../src/status/service'
 import { ConduitStatus } from '../../src/status/types'
-import { app, randStr, setupApp } from './utils'
+import { app, randStr, setupApp } from '../utils'
 
 const TOLERANCE = ms('1h')
 const LIMIT = 10
